+++ core-ng/src/test/java/core/framework/impl/web/management/ThreadInfoControllerTest.java
@@ -3,8 +3,7 @@ package core.framework.impl.web.management;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
+import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * @author neo
@@ -20,6 +19,6 @@ class ThreadInfoControllerTest {
     @Test
     void threadDump() {
         String threadDump = controller.threadDumpText();
-        assertThat(threadDump, containsString("core.framework.impl.web.management.ThreadInfoController.threadDumpText"));
+        assertThat(threadDump).contains("core.framework.impl.web.management.ThreadInfoController.threadDumpText");
     }
 }