+++ core-ng/src/test/java/core/framework/impl/web/request/PathParamsTest.java
@@ -4,8 +4,7 @@ import core.framework.web.exception.BadRequestException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 /**
@@ -22,6 +21,6 @@ class PathParamsTest {
     @Test
     void putEmptyPathParam() {
         BadRequestException exception = assertThrows(BadRequestException.class, () -> pathParams.put("id", ""));
-        assertThat(exception.getMessage(), containsString("name=id, value="));
+        assertThat(exception.getMessage()).contains("name=id, value=");
     }
 }