+++ core-ng/src/test/java/core/framework/util/PropertiesTest.java
@@ -3,8 +3,7 @@ package core.framework.util;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.containsString;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -29,6 +28,6 @@ class PropertiesTest {
     @Test
     void loadNotExistedFile() {
         Error error = assertThrows(Error.class, () -> properties.load("not-existed-property.properties"));
-        assertThat(error.getMessage(), containsString("can not find"));
+        assertThat(error.getMessage()).contains("can not find");
     }
 }