+++ core-ng/src/test/java/core/framework/util/RandomsTest.java
@@ -2,8 +2,7 @@ package core.framework.util;
 
 import org.junit.jupiter.api.Test;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 
@@ -13,9 +12,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 class RandomsTest {
     @Test
     void randomAlphaNumeric() {
-        assertThat(Randoms.alphaNumeric(3).length(), equalTo(3));
-        assertThat(Randoms.alphaNumeric(5).length(), equalTo(5));
-        assertThat(Randoms.alphaNumeric(10).length(), equalTo(10));
+        assertThat(Randoms.alphaNumeric(3).length()).isEqualTo(3);
+        assertThat(Randoms.alphaNumeric(5).length()).isEqualTo(5);
+        assertThat(Randoms.alphaNumeric(10).length()).isEqualTo(10);
     }
 
     @Test